+++ crates/rattler/src/libsolv/ffi.rs
