+++ crates/rattler/src/main.rs
@@ -1,5 +1,3 @@
-extern crate core;
-
 use clap::Parser;
 
 mod commands;
